fix warnings and errors found by PVS
..somebody please review this

git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3610 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/sound/adpcm.c
@@ -769,7 +769,7 @@ void OKIM6295_set_bank_base(int which, int base)
 
 ***********************************************************************************************/
 
-void OKIM6295_set_frequency(int which, int frequency)
+void OKIM6295_set_frequency(int which, double frequency)
 {
 	int channel;
 
@@ -780,7 +780,7 @@ void OKIM6295_set_frequency(int which, int frequency)
 		/* update the stream and set the new base */
 		stream_update(voice->stream, 0);
 		if (Machine->sample_rate)
-			voice->source_step = (UINT32)((double)frequency * (double)FRAC_ONE / (double)Machine->sample_rate);
+			voice->source_step = (UINT32)(frequency * (double)FRAC_ONE / (double)Machine->sample_rate);
 	}
 }
